+++ src/components/ConjectureModule/ConjectureModule.js
@@ -5,7 +5,7 @@ import RectButton from "../RectButton";
 import InputBox from "../InputBox";
 import { ConjectureBox, KeywordsBox, NameBox, PINBox } from "./ConjectureModuleBoxes";
 import { EndBox, IntermediateBox, StartBox } from "../PoseAuth/PoseAuthoringBoxes";
-import { writeToDatabaseConjecture, writeToDatabaseDraft, getConjectureDataByUUID } from "../../firebase/database";
+import { writeToDatabaseConjecture, writeToDatabaseDraft, getConjectureDataByUUID,getConjectureDataByAuthorID, getConjectureDataByPIN } from "../../firebase/database";
 import { useMachine } from "@xstate/react";
 import { ConjectureEditorMachine } from "../../machines/conjectureEditorMachine";
 
@@ -130,11 +130,56 @@ const ConjectureModule = (props) => {
           color={black}
           fontSize={width * 0.015}
           fontColor={white}
-          text={"TEST BUTTON"}
+          text={"TEST UUID"}
           fontWeight={800}
           callback={() => {
-            const targetUUID = "3922e862-a36c-445b-a186-ccf170c6b967"; // a UUID I created earlier
+            const targetUUID = "698c9d7d-2d4a-44e4-8ac1-abee27569a6b"; // a UUID I created earlier
             getConjectureDataByUUID(targetUUID)
+              .then((conjectureData) => {
+                console.log('Conjecture Data:', conjectureData);
+              })
+              .catch((error) => {
+                console.error('Error getting conjecture data: ', error);
+              });
+          }}
+        />
+        {/* TEST Author ID Button */}
+        <RectButton
+          height={height * 0.13}
+          width={width * 0.26}
+          x={width * 0.05}
+          y={height * 0.93}
+          color={black}
+          fontSize={width * 0.015}
+          fontColor={white}
+          text={"TEST AuthorID"}
+          fontWeight={800}
+          callback={() => {
+            const userId = "CYONFYmwvXbPT1uLUqkTO2GPtoL2"; // nates user id
+            getConjectureDataByAuthorID(userId)
+              .then((conjectureData) => {
+                console.log('Conjecture Data:', conjectureData);
+              })
+              .catch((error) => {
+                console.error('Error getting conjecture data: ', error);
+              });
+          }}
+        />
+
+        {/* TEST PIN Button */}
+        <RectButton
+          height={height * 0.13}
+          width={width * 0.26}
+          x={width * 0.25}
+          y={height * 0.93}
+          color={black}
+          fontSize={width * 0.015}
+          fontColor={white}
+          text={"TEST PIN"}
+          fontWeight={800}
+          callback={() => {
+            const PIN = "123456789"; // some PIN I created
+            getConjectureDataByPIN(PIN)
               .then((conjectureData) => {
                 console.log('Conjecture Data:', conjectureData);
               })